Rename event file to skateevent
